ai-news-backend/requirements.txt
flask>=2.0.0
requests>=2.28.0
feedparser>=6.0.0
gunicorn>=20.0.0
Pillow>=10.0.0
playwright>=1.40.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
